backend/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Bcrypt>=1.0
Flask-JWT-Extended>=4.6
Flask-Cors>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
pdfplumber>=0.11
numpy>=1.26
transformers>=4.46
sentence-transformers>=3.1
langchain>=0.3
langchain-community>=0.3
langchain-core>=0.3
faiss-cpu>=1.8
torch>=2.5
pandas>=2.2

# PDF Processing
pdfplumber>=0.11.0

# Machine Learning & Transformers
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=5.0.0
huggingface-hub>=0.20.0

# Vector Search & Embeddings
faiss-cpu>=1.8.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# LangChain Framework
langchain>=0.3.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-text-splitters>=1.0.0

# Graph & Network Analysis
networkx>=3.0

# Visualization
matplotlib>=3.7.0
pillow>=9.0.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For GPU support (uncomment if using CUDA)
# faiss-gpu>=1.8.0
# torch>=2.0.0+cu118

# Development
setuptools>=65.0.0
